refactor: rename modules of board column entities
