+++ src/resources/tasks/task.types.ts
@@ -6,7 +6,7 @@ import { genId } from '../../common/utils';
 
 import { UserId } from '../users/user.types';
 import { BoardId, IBoardId } from '../boards/board.types';
-import { BoardColumnId } from '../boards/column.types';
+import { BoardColumnId } from '../boards/board-column.types';
 
 /**
  * The unique **Id** of the task.
